finalizes blog, refactors directories
