--- GGCHacks/src/app/blog-home/blog-home.component.html
@@ -1,32 +1,0 @@
-<div class="everything">
-    <div *ngIf="blogs === null || blogs.length === 0">
-        <div class="card w-50 h-80">Loading...</div>
-    </div>
-    <div *ngIf="blogs && blogs.length !== 0">
-        <div *ngFor="let blog of blogs">
-            <div class="row justify-content-center">
-                <div class="card w-50 h-80">
-                    <div class="row justify-content-center">
-                        <h1 class="justify-content-center">{{blog.title}}</h1>
-                    </div>
-                    <div class="authorDateRow row">
-                        <div class="col d-flex align-items-center"></div>
-                        <div class="col-auto d-flex align-items-center">
-                            <h6 class="content">Author: {{blog.author}}, {{blog.datePosted.toDate().toDateString()}}</h6>
-                        </div>
-                        <div class="col d-flex align-items-right">
-                            <ng-container *ngIf="auth?.user?.isSuperAdmin">
-                                <button class="btn btn-primary" (click)="editBlog(blog)">Edit</button>
-                                <button class="btn btn-primary" (click)="deleteBlog(blog)">Delete</button>
-                            </ng-container>
-                        </div>
-                    </div>
-                    <hr/>
-                    <div id="blog-content" class="row justify-content-left">
-                        <p class="blog-content">{{blog.content}}</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
