+++ GGCHacks/src/app/blog/blog-home/edit-blog-modal/edit-blog-modal.component.css
@@ -1,0 +1,3 @@
+header {
+    padding: 50px;
+}
