+++ GGCHacks/src/app/blog/blog-home/edit-blog-modal/edit-blog-modal.component.html
@@ -11,6 +11,10 @@
       <input type="text" class="form-control" id="title" formControlName="title">
     </div>
     <div class="form-group">
+      <label for="author">Author</label>
+      <input type="text" class="form-control" id="author" formControlName="author">
+    </div>
+    <div class="form-group">
       <label for="content">Content</label>
       <textarea class="form-control" rows="10" id="content" formControlName="content"></textarea>
     </div>
