+++ GGCHacks/src/app/header/header.component.html
@@ -7,7 +7,6 @@
   </a>
   </div>
   <div class="justify-content-end">
-    <button class="btn btn-primary" routerLink="/blogPostingPage">Post Blog</button>
 
     <button class="btn btn-sm btn-outline-secondary" type="button"  routerLink="/home">Home</button>
