+++ GGCHacks/src/styles.css
@@ -74,7 +74,8 @@ html
   linear-gradient(rgb(0,82,150), rgb(20,25,58)) !important;
   background-size: cover !important;
   margin: 0 !important;
-  padding-bottom: 100px;
+  min-height: 100vh !important;
+  padding-bottom: 50px !important;
 }
 
 header{
